docker/print_page/requirements.txt
numpy>=1.14.3
Pillow>=5.0.0
PyMuPDF==1.14.21
pytesseract>=0.3.4
opencv-python

